+++ cmd/account/billable-usage/billable-usage.go
@@ -72,6 +72,9 @@ var downloadCmd = &cobra.Command{
 		}
 		return nil
 	},
+	// Disable completions since they are not applicable.
+	// Can be overridden by manual implementation in `override.go`.
+	ValidArgsFunction: cobra.NoFileCompletions,
 }
 
 // end service BillableUsage