+++ cmd/workspace/system-schemas/system-schemas.go
@@ -19,6 +19,9 @@ var Cmd = &cobra.Command{
 	Annotations: map[string]string{
 		"package": "catalog",
 	},
+
+	// This service is being previewed; hide from help output.
+	Hidden: true,
 }
 
 // start disable command
@@ -69,6 +72,9 @@ var disableCmd = &cobra.Command{
 		}
 		return nil
 	},
+	// Disable completions since they are not applicable.
+	// Can be overridden by manual implementation in `override.go`.
+	ValidArgsFunction: cobra.NoFileCompletions,
 }
 
 // start enable command
@@ -97,6 +103,9 @@ var enableCmd = &cobra.Command{
 		}
 		return nil
 	},
+	// Disable completions since they are not applicable.
+	// Can be overridden by manual implementation in `override.go`.
+	ValidArgsFunction: cobra.NoFileCompletions,
 }
 
 // start list command
@@ -146,6 +155,9 @@ var listCmd = &cobra.Command{
 		}
 		return cmdio.Render(ctx, response)
 	},
+	// Disable completions since they are not applicable.
+	// Can be overridden by manual implementation in `override.go`.
+	ValidArgsFunction: cobra.NoFileCompletions,
 }
 
 // end service SystemSchemas